[LR11x0] Fix SNR calculation (#1161)

+++ src/modules/LR11x0/LR11x0.cpp
@@ -2374,7 +2374,7 @@ int16_t LR11x0::getPacketStatusLoRa(float* rssiPkt, float* snrPkt, float* signal
 
   // pass the replies
   if(rssiPkt) { *rssiPkt = (float)buff[0] / -2.0f; }
-  if(snrPkt) { *snrPkt = (float)buff[1] / 4.0f; }
+  if(snrPkt) { *snrPkt = (float)((int8_t)buff[1]) / 4.0f; }
   if(signalRssiPkt) { *signalRssiPkt = buff[2]; }
 
   return(state);
